requirements.txt
Django==3.1.1
django-auto-logout==0.5.1
pytz==2023.3
mysqlclient==2.1.0
django-mysql==3.8.1
werkzeug==2.2.2
Pillow==9.2.0
gunicorn==20.1.0
psycopg2-binary==2.9.1
python-telegram-bot==12.2.0
tornado==5.1.1
requests==2.7.0
python-crontab==2.7.1
celery==5.2.7
redis==4.5.5
flower==1.2.0
turbosmsua==0.3
twilio==8.2.1
celery==5.2.7
redis==4.5.5
liqpay-sdk-python3==1.0.6
openpyxl==3.1.2
aiogram==3.4.1